refactor: move .js files to lib directory
BREAKING CHANGE: External dependencies may need to be updated
