+++ lib/asyncapi1.js
@@ -82,7 +82,7 @@ function convertInner(api, options, callback) {
     header.highlight_theme = options.theme;
 
     if (typeof templates === 'undefined') {
-        templates = dot.process({ path: path.join(__dirname, 'templates', 'asyncapi1') });
+        templates = dot.process({ path: path.join(__dirname, '..', 'templates', 'asyncapi1') });
     }
     if (options.user_templates) {
         templates = Object.assign(templates, dot.process({ path: options.user_templates }));
