+++ lib/common.js
@@ -13,7 +13,7 @@ const httpsnippetGenerator = require('./httpsnippetGenerator');
 
 /* originally from https://github.com/for-GET/know-your-http-well/blob/master/json/status-codes.json */
 /* "Unlicensed", public domain */
-const statusCodes = require('./statusCodes.json');
+const statusCodes = require('..//statusCodes.json');
 
 const contentTypes = {
     xml: ['^(application|text|image){1}\\/(.*\\+){0,1}xml(;){0,1}(\\s){0,}(charset=.*){0,}$'],
