+++ lib/harGenerator.js
