+++ lib/httpsnippetGenerator.js
