+++ lib/openapi3.js
@@ -600,7 +600,7 @@ function convertInner(api, options, callback) {
     }
 
     if (typeof templates === 'undefined') {
-        templates = dot.process({ path: path.join(__dirname, 'templates', 'openapi3') });
+        templates = dot.process({ path: path.join(__dirname, '..', 'templates', 'openapi3') });
     }
     if (options.user_templates) {
         templates = Object.assign(templates, dot.process({ path: options.user_templates }));
@@ -666,7 +666,7 @@ function convertInner(api, options, callback) {
     data.utils.getResponseExamples = getResponseExamples;
     data.utils.getResponseHeaders = getResponseHeaders;
     data.utils.getAuthenticationStr = getAuthenticationStr;
-    data.widdershins = require('./package.json');
+    data.widdershins = require('../package.json');
     data.utils.join = function(s) {
         return s.split('\r').join('').split('\n').join(' ').trim();
     };
