+++ lib/semoasa.js
@@ -9,7 +9,7 @@ dot.templateSettings.varname = 'data';
 
 const common = require('./common.js');
 const dereference = require('reftools/lib/dereference.js').dereference;
-const oas_descs = require('./resources/oas_descs.js');
+const oas_descs = require('../resources/oas_descs.js');
 
 let templates = {};
 
@@ -44,7 +44,7 @@ function convert(api, options, callback) {
     header.highlight_theme = options.theme;
 
     if (typeof templates === 'undefined') {
-        templates = dot.process({ path: path.join(__dirname, 'templates', 'semoasa') });
+        templates = dot.process({ path: path.join(__dirname, '..', 'templates', 'semoasa') });
     }
     if (options.user_templates) {
         templates = Object.assign(templates, dot.process({ path: options.user_templates }));
